phishguard-ai/requirements.txt
# PhishGuard AI - Requirements
# Install: pip install -r requirements.txt

# Web Framework
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0

# Authentication
werkzeug==3.0.1
bcrypt==4.1.2

# Machine Learning
scikit-learn==1.4.0
numpy==1.26.3
scipy==1.12.0

# NLP
transformers==4.37.2
torch==2.2.0
tokenizers==0.15.1

# Database
sqlalchemy==2.0.25

# HTTP & Networking
requests==2.31.0

# WHOIS lookup (optional - for deep mode)
python-whois==0.9.4

# PDF generation (optional - for reports)
reportlab==4.1.0

# Utilities
python-dotenv==1.0.1
